[EDT-1538] feat: unit tests
